renamed to navigator_server & cpplint passed
