+++ nav2_bt_navigator/src/main.cpp
@@ -14,13 +14,13 @@
 
 #include <memory>
 
-#include "nav2_bt_navigator/bt_navigator.hpp"
+#include "nav2_bt_navigator/navigator_server.hpp"
 #include "rclcpp/rclcpp.hpp"
 
 int main(int argc, char ** argv)
 {
   rclcpp::init(argc, argv);
-  auto node = std::make_shared<nav2_bt_navigator::BtNavigator>();
+  auto node = std::make_shared<nav2_bt_navigator::NavigatorServer>();
   rclcpp::spin(node->get_node_base_interface());
   rclcpp::shutdown();
